requirements.txt
Keras==2.4.3
numpy==1.19.5
tensorflow-gpu==2.4.0
